requirements.txt
langchain==0.1.14
langchain-community==0.0.34
langchain-core==0.1.45
langchain-text-splitters==0.0.1
openrouter==1.0.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
tiktoken==0.7.0
requests==2.31.0
streamlit==1.32.0